Framework for ML/requirements.txt
appdirs==1.4.4
certifi==2020.12.5
click==7.1.2
DateTime==4.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
imutils==0.5.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
pipenv==2020.11.15
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
virtualenv==20.4.1
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zope.interface==5.2.0
opencv-python-headless==4.5.1.48
opencv-contrib-python-headless==4.5.1.48